refactor: 도메인 구조 분리 (point, payment, virtual-account)
