+++ src/main/java/com/biddingmate/biddinggo/payment/mapper/PaymentMybatisMapper.java
@@ -1,7 +1,7 @@
-package com.biddingmate.biddinggo.point.mapper;
+package com.biddingmate.biddinggo.payment.mapper;
 
 import com.biddingmate.biddinggo.common.inif.IMybatisCRUD;
-import com.biddingmate.biddinggo.point.model.Payment;
+import com.biddingmate.biddinggo.payment.model.Payment;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
