+++ src/main/java/com/biddingmate/biddinggo/virtualaccount/model/VirtualAccount.java
@@ -1,20 +1,24 @@
-package com.biddingmate.biddinggo.point.dto;
+package com.biddingmate.biddinggo.virtualaccount.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.time.LocalDateTime;
 
 @Getter
+@Setter
 @AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class CreateVirtualAccountResponse {
-    private String status;
-    private String orderId;
-    private Long amount;
+public class VirtualAccount {
+    private Long id;
+    private Long paymentId;
     private String bankCode;
     private String bankAccount;
     private String accountHolderName;
     private LocalDateTime dueDate;
+    private LocalDateTime createdAt;
 }
