+++ src/main/java/com/biddingmate/biddinggo/virtualaccount/service/VirtualAccountService.java
@@ -1,0 +1,8 @@
+package com.biddingmate.biddinggo.virtualaccount.service;
+
+import com.biddingmate.biddinggo.virtualaccount.dto.CreateVirtualAccountRequest;
+import com.biddingmate.biddinggo.virtualaccount.dto.CreateVirtualAccountResponse;
+
+public interface VirtualAccountService {
+    CreateVirtualAccountResponse createVirtualAccount(CreateVirtualAccountRequest request);
+}
