+++ src/main/resources/mapper/payment/PaymentMybatisMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.biddingmate.biddinggo.point.mapper.PaymentMybatisMapper">
+<mapper namespace="com.biddingmate.biddinggo.payment.mapper.PaymentMybatisMapper">
 
     <insert id="insert" parameterType="Payment" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO payment(
